pyproject.toml
[project]
name = "black-stories-ai"
version = "0.1.0"
description = "Black Stories game with AI models competing"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.scripts]
blackstory = "main:main"
